refactor: register profile tab with donor tabs register
